+++ app/Http/Controllers/InvoicesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Exports\InvoiceExport;
+use App\Http\Requests\InvoiceRequest;
 use App\Models\Section;
 use App\Models\invoices;
 use Illuminate\Http\Request;
@@ -51,17 +52,19 @@ class invoicesController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(InvoiceRequest $request)
     {
+        $validate = $request->validate();
+
         invoices::create([
             'invoice_number' => $request->invoice_number,
-            'invoice_date' => $request->invoice_Date,
-            'due_date' => $request->Due_date,
+            'invoice_date' => $request->invoice_date,
+            'due_date' => $request->due_date,
             'branch_id' => $request->branch,
-            'section_id' => $request->Section,
+            'section_id' => $request->section,
             'amount_collection' => $request->amount_collection,
             'amount_commission' => $request->amount_commission,
             'discount' => $request->discount,
@@ -82,7 +85,6 @@ class invoicesController extends Controller
         ]);
 
 
-
         if ($request->hasFile('pic')) {
             $request->invoice_id = $invoice_id;
             $this->addAttachment($request);
@@ -100,7 +102,7 @@ class invoicesController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\invoices  $invoices
+     * @param \App\Models\invoices $invoices
      * @return \Illuminate\Http\Response
      */
     public function show(invoices $invoices)
@@ -111,32 +113,34 @@ class invoicesController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\invoices  $invoices
+     * @param \App\Models\invoices $invoices
      * @return \Illuminate\Http\Response
      */
-    public function edit($invoices)
+    public function edit()
     {
-        return $invoices;
+
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\invoices  $invoices
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Models\invoices $invoices
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request)
+    public function update(InvoiceRequest $request)
     {
+        $validate = $request->validate();
+
         // for get old invoice number and get new from request
         $invoice = invoices::find($request->id);
 
         invoices::where('id', $request->id)->update([
             'invoice_number' => $request->invoice_number,
-            'invoice_date' => $request->invoice_Date,
-            'due_date' => $request->Due_date,
+            'invoice_date' => $request->invoice_date,
+            'due_date' => $request->due_date,
             'branch_id' => $request->branch,
-            'section_id' => $request->Section,
+            'section_id' => $request->section,
             'amount_collection' => $request->amount_collection,
             'amount_commission' => $request->amount_commission,
             'discount' => $request->discount,
@@ -148,9 +152,8 @@ class invoicesController extends Controller
         ]);
 
 
-
         if (isset($invoice->invoices_attachments)) {
-            rename(public_path('Attachments/' . $invoice->invoice_number),  public_path('Attachments/' . $request->invoice_number));
+            rename(public_path('Attachments/' . $invoice->invoice_number), public_path('Attachments/' . $request->invoice_number));
             // Storage::disk('public_uploads')->deleteDirectory($invoice->invoice_number);
         }
 
@@ -161,34 +164,38 @@ class invoicesController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\invoices  $invoices
+     * @param \App\Models\invoices $invoices
      * @return \Illuminate\Http\Response
      */
     public function destroy(Request $request)
     {
         $id = $request->invoice_id;
         $invoice = invoices::where('id', $id)->first();
-        $invoice->delete();
-        session()->flash('archived_invoice');
+
+        if (isset($invoice)) {
+            $invoice->delete();
+            session()->flash('archived_invoice');
+        }
+
         return back();
     }
-
 
 
     public function addAttachment(Request $request)
     {
 
+        $validate = $request->validate([
+            'pic' => 'required|max:255|mimes:pdf,jpeg,png,jpg',
+            'invoice_number' => 'required|max:255',
+        ]);
+
         if ($request->hasFile('pic')) {
 
-            $request->validate([
-                'pic' => 'mimes:pdf,jpeg,png,jpg',
-            ]);
-
-            $fileName =  $request->file('pic')->getClientOriginalName();
-            $invoice_number =  $request->invoice_number;
+            $fileName = $request->file('pic')->getClientOriginalName();
+            $invoice_number = $request->invoice_number;
             invoices_attachment::create(
                 [
-                    'file_name' =>  $fileName,
+                    'file_name' => $fileName,
                     'user_id' => Auth::user()->id,
                     'invoices_id' => $request->invoice_id
                 ]
@@ -196,29 +203,39 @@ class invoicesController extends Controller
 
             $request->pic->move(public_path('Attachments/' . $invoice_number), $fileName);
         }
+
         return back();
     }
 
 
     public function updateDetails(Request $request)
     {
-
-        invoices_details::create([
-            'invoices_id' => $request->id,
-            'amount_paid' => $request->amount_paid,
-            'note' => $request->note,
-            'user_id' => (Auth::user()->id),
+        $validate = $request->validate([
+            'id' => 'required',
+            'amount_paid' => 'required',
+            'note' => 'max:255',
         ]);
 
         $invpice = invoices::find($request->id);
-        $invpice->total_paid += $request->amount_paid;
 
-        if ($invpice->total_paid >= $invpice->total_amount) {
-            $invpice->value_status = 2;
-        } else if ($invpice->total_paid > 0) {
-            $invpice->value_status = 1;
+        if (isset($invpice)) {
+
+            invoices_details::create([
+                'invoices_id' => $request->id,
+                'amount_paid' => $request->amount_paid,
+                'note' => $request->note,
+                'user_id' => (Auth::user()->id),
+            ]);
+
+            $invpice->total_paid += $request->amount_paid;
+
+            if ($invpice->total_paid >= $invpice->total_amount) {
+                $invpice->value_status = 2;
+            } else if ($invpice->total_paid > 0) {
+                $invpice->value_status = 1;
+            }
+            $invpice->save();
         }
-        $invpice->save();
         return $this->index();
     }
 
@@ -227,57 +244,86 @@ class invoicesController extends Controller
         $invoices = invoices::where('value_status', 2)->get();
         return view('invoices.invoices_paid', compact('invoices'));
     }
+
     public function invoices_unpaid()
     {
         $invoices = invoices::where('value_status', 0)->get();
         return view('invoices.invoices_unpaid', compact('invoices'));
     }
+
     public function invoices_partially_paid()
     {
         $invoices = invoices::where('value_status', 1)->get();
         return view('invoices.invoices_partially_paid', compact('invoices'));
     }
+
     public function deleted_invoiced()
     {
         $invoices = invoices::onlyTrashed()->get();
         return view('invoices.deleted_invoiced', compact('invoices'));
     }
+
     public function archived_invoiced()
     {
         $invoices = invoices::where('archive', '=', 1)->get();
         // return $invoices;
         return view('invoices.archived_invoiced', compact('invoices'));
     }
+
     public function restore(Request $request)
     {
-        $id = $request->invoice_id;
-        $invoice = invoices::withTrashed()->where('id', $id)->restore();
-        session()->flash('restore');
-        return back();
-    }
-    public function destroyWithTrashed(Request $request)
-    {
+        $validate = $request->validate([
+            'invoice_id' => 'required',
+        ]);
+
         $id = $request->invoice_id;
         $invoice = invoices::withTrashed()->where('id', $id)->first();
 
-        if (isset($invoice->invoices_attachments)) {
-            Storage::disk('public_uploads')->deleteDirectory($invoice->invoice_number);
+        if (isset($invoice)) {
+            $invoice = $invoice->restore();
+            session()->flash('restore');
         }
-
-        $invoice->forceDelete();
-        session()->flash('delete_invoice');
         return back();
     }
+
+    public function destroyWithTrashed(Request $request)
+    {
+        $validate = $request->validate([
+            'invoice_id' => 'required',
+        ]);
+
+        $id = $request->invoice_id;
+        $invoice = invoices::withTrashed()->where('id', $id)->first();
+
+        if (isset($invoice)) {
+            if (isset($invoice->invoices_attachments)) {
+                Storage::disk('public_uploads')->deleteDirectory($invoice->invoice_number);
+            }
+
+            $invoice->forceDelete();
+            session()->flash('delete_invoice');
+        }
+
+        return back();
+    }
+
     public function print_invoice($id)
     {
 
         $invoice = invoices::find($id);
-        return view('invoices.print_invoice', compact('invoice', 'details'));
+
+        if (isset($invoice)) {
+            return view('invoices.print_invoice', compact('invoice', 'details'));
+        } else {
+            return back();
+        }
     }
+
     public function export()
     {
         return Excel::download(new InvoiceExport, 'Invoices.xlsx');
     }
+
     public function archive(Request $request)
     {
         if ($request->page_id == 2) {
@@ -288,6 +334,7 @@ class invoicesController extends Controller
 
         return back();
     }
+
     public function markAsRead()
     {
         auth()->user()->unreadNotifications->markAsRead();
@@ -299,13 +346,13 @@ class invoicesController extends Controller
     public function search($txt)
     {
         $result = invoices::search($txt)
-        ->with(['invoices_attachments' , 'invoices_details' ])
-        ->withTrashed()
-        ->distinct()
-        ->get();
+            ->with(['invoices_attachments', 'invoices_details'])
+            ->withTrashed()
+            ->distinct()
+            ->get();
 
-        $result= isset( $result) ?  $result : ["found" => "not result found"];
+        $result = isset($result) ? $result : ["found" => "not result found"];
 
-        return  response()->json($result);
+        return response()->json($result);
     }
 }
